// frontend/front_pfe_standalone-houssem/src/app/services/models/backlog-item-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BacklogItemResponse {
  completed?: boolean;
  description?: string;
  id?: number;
  inProgress?: boolean;
  points?: number;
  priorite?: number;
  productBacklogId?: number;
  productBacklogTitre?: string;
  projetNom?: string;
  statut?: 'EN_COURS' | 'TERMINE' | 'A_FAIRE';
  statutLabel?: string;
  titre?: string;
  type?: 'USER_STORY' | 'BUG' | 'TECHNICAL_TASK';
}