// frontend/front_pfe_standalone-houssem/src/app/services/models/backlog-item-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BacklogItemRequest {
  description?: string;
  points?: number;
  priorite: number;
  productBacklogId: number;
  statut?: 'EN_COURS' | 'TERMINE' | 'A_FAIRE';
  titre: string;
  type?: 'USER_STORY' | 'BUG' | 'TECHNICAL_TASK';
}